src/function/simulation_publish_messages/requirements.txt
functions-framework
google-cloud-bigquery
google-cloud-pubsub